backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic
google-genai
pdfminer.six
python-pptx
python-multipart